Plot caterpillar chart

this function plots a caterpillar chart using the dataframe created by create_caterpillar_df()

plot_caterpillar_chart(
  df,
  y_var = rate,
  x_var = site_name,
  point_shape = "diamond",
  point_size = 2,
  bar_width = 3,
  dt_color = OBI.color::prim_dark_blue()
)

Arguments

df

Data frame created by create_caterpillar_df()

y_var

Variable to be placed on the y axis - default is rate, which is the rate variable created by create_caterpillar_df()

x_var

Variable to be placed on the x axis - default is site_name

point_shape

Shape of the points - default is "diamond"

point_size

Size of the points - default is 2

bar_width

Width of the error bars - default is 3

dt_color

Color for the data points and errorbars - default is OBI.color::prim_dark_blue()